Updated “WebFont”. The last updated reported a server 500 error, so I am trying again, though I have added a few more improvements.
Last update that failed:
Adjusted the small k(s) so that the secondary arm is under x height. Adjusted all of the periods and dots to be the same, adjusted the commas, apostrophes and quotes to all be the same. This version should show more uniformity and compliance with SIL character design directives, especially as I corrected the cedillas and carons that are actually looking more like commas.

Since then, a fontstruct designer, @TCWhite (on fontsruct) spotted a few more periods, dots, commas, quotes I needed to address and did so.

There were a few stray lowercase letters that didn't have goth cuts on the top acsender.  I can't add got cuts to the x, unfortunately.  That is the issue with the limited choice of bricks we have to build a font (I keep asking for more circles, especially, because the curve of a P, R, S is tighter than what's required for an O, D, C, etc.).

I also adjusted some of the spacing for dingbat characters and the like.

Here is what is left on my list to do...

There may be some issues I have to address...

1) Lookup what a Latin lowercase 6 is & lookup what a Cyrillic small letter soft sign is.  Then fix the issue.

2) Lookup what a schwa is. Then fix the issue.

3) Find #0259 character and fix it.

4) Lookup what a swing dash is and fix it.

5) Lookup what a commercial minus is and fix it.

Sorry that this update took a while.  I do this in my spare time and I have way too many irons in the fire.

I'm still willing to listen to any criticism, but I would like to get this one accepted (once it is what is expected) because I have a whole font family awaiting further development based on this model.

Thank you!