Added “Middags”. A chunky rounded display font offering two different visuals: upper case has solid black chunky lines, lower case has these lines split into vertical 'slices' making LC wider than UC.
Middag's rounded chunky shapes are a little unusual which should attract viewers' eyes when it is used for printed or painted signs for shops or vehicles, logos, splash words on a leaflet/magazine, short headlines and messages. It could provide surprise and maybe entertainment if UC glyphs were painted or printed on vertical or horizontal 'venetian blinds'.

LC glyphs look a little more open to observation and might invite 'investigation' keeping viewer concentration on the text: ideal for logos, signs, single words, short phrases and messages. Many LC glyphs look good when used as graphic elements to embellish pages, margins, plain texts, or as dividers.

Middag could be used in a 3D presentation, to attract and maintain attention, by constructing LC glyphs with a certain volume and, for example, using them to make words with glyphs standing or lying in windows or hanging from objects -like ceilings, masts etc- to give eye catching 3D messages.

The original version of this design was inspired by slices of a tree trunk which showed split branchlets and the bases of branches emerging as nodules which added unexpected lines and those well-known knots to the internal patterning of the tree slices.